deposit overview class renames
