+++ src/main/typescript/components/overview/DepositOverview.tsx
@@ -18,7 +18,7 @@ import { Component } from "react"
 import { connect } from "react-redux"
 import { AppState } from "../../model/AppState"
 import { DepositId, DepositOverviewState } from "../../model/Deposits"
-import { cleanDeposits, deleteDeposit, fetchDeposits } from "../../actions/depositActions"
+import { cleanDeposits, deleteDeposit, fetchDeposits } from "../../actions/depositOverviewActions"
 import { ReduxAction } from "../../lib/redux"
 import { Action } from "redux"
 import DepositTableHead from "./DepositTableHead"
