+++ src/main/typescript/middleware/rejectedRequestMiddleware.ts
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { DepositConstants } from "../constants/depositConstants"
+import { DepositOverviewConstants } from "../constants/depositOverviewConstants"
 import { Action, AnyAction, Middleware } from "redux"
 import { createMiddleware } from "../lib/redux"
-import { deleteDepositFailed, fetchDepositsFailed } from "../actions/depositActions"
+import { deleteDepositFailed, fetchDepositsFailed } from "../actions/depositOverviewActions"
 
 type NewActionGenerator = (action: AnyAction) => (errorMessage: string) => Action
 
@@ -37,8 +37,8 @@ function rejectedMiddleware(type: string) {
     }
 }
 
-const depositFetchRejected = rejectedMiddleware(DepositConstants.FETCH_DEPOSITS_REJECTED)(() => fetchDepositsFailed)
+const depositFetchRejected = rejectedMiddleware(DepositOverviewConstants.FETCH_DEPOSITS_REJECTED)(() => fetchDepositsFailed)
 
-const depositDeleteRejected = rejectedMiddleware(DepositConstants.DELETE_DEPOSIT_REJECTED)(({ meta: { depositId } }) => deleteDepositFailed(depositId))
+const depositDeleteRejected = rejectedMiddleware(DepositOverviewConstants.DELETE_DEPOSIT_REJECTED)(({ meta: { depositId } }) => deleteDepositFailed(depositId))
 
 export const rejectedRequestMiddleware = [depositFetchRejected, depositDeleteRejected]
